+++ Content/Items/Dedicated/Fluffy/GiantBastStatue.cs
@@ -8,6 +8,7 @@ using Terraria.Audio;
 using Terraria.ID;
 using Terraria.ModLoader;
 using TwilightEgress.Core;
+using TwilightEgress.Core.Graphics;
 
 namespace TwilightEgress.Content.Items.Dedicated.Fluffy
 {
