+++ Content/Items/Dedicated/Fluffy/HomingBastStatue.cs
@@ -11,6 +11,7 @@ using Terraria.ModLoader;
 using TwilightEgress.Content.Particles;
 using TwilightEgress.Core;
 using TwilightEgress.Core.Globals.GlobalProjectiles;
+using TwilightEgress.Core.Graphics;
 
 namespace TwilightEgress.Content.Items.Dedicated.Fluffy
 {
