+++ Content/Items/Dedicated/Marv/ElectricSkyBoltExplosion.cs
@@ -9,6 +9,7 @@ using Terraria.ModLoader;
 using TwilightEgress.Assets;
 using TwilightEgress.Content.Particles;
 using TwilightEgress.Core;
+using TwilightEgress.Core.Graphics;
 
 namespace TwilightEgress.Content.Items.Dedicated.Marv
 {
