+++ Content/Items/FrostMoon/HolidayHalberdIceShock.cs
@@ -12,6 +12,7 @@ using TwilightEgress.Assets;
 using TwilightEgress.Content.Particles;
 using TwilightEgress.Core;
 using TwilightEgress.Core.Globals.GlobalProjectiles;
+using TwilightEgress.Core.Graphics;
 
 namespace TwilightEgress.Content.Items.FrostMoon
 {
