+++ Content/Items/FrostMoon/HolidayHalberdIceShockSnowflake.cs
@@ -7,7 +7,7 @@ using Terraria.Audio;
 using Terraria.DataStructures;
 using Terraria.ID;
 using Terraria.ModLoader;
-using TwilightEgress.Core;
+using TwilightEgress.Core.Graphics;
 
 namespace TwilightEgress.Content.Items.FrostMoon
 {
